requirements.txt
fredapi==0.5.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
streamlit>=1.54.0
plotly>=6.5.0
Jinja2>=3.1.0
matplotlib>=3.8.0
requests>=2.31.0
pyyaml>=6.0
